refactor(repository): flatten repository package hierarchy
Remove unnecessary package nesting for repository interfaces. Group related repositories under a cleaner structure to reduce directory depth and improve maintainability.
